Improve parsing of dependancies and fix error output.
Improve the parsing of dependancies by making it more robust.  It
is now tolerant of ports that omit $PORTSDIR from the port path.

With the change to stderr piping, only the stderr will be reported if
make fails, also if stderr output is given then a "warning" message
will be recorded.  Lastly, log errors produced by the attribute
parsing functions.