Remove environ variables used by (p)make and set MACHTYPE for --arch.
Make sets various environment variables that may interfer when
portbuilder is run from a Makefile.  This protects from such
interference, but is dependant on internal make interfaces and may
break with changes to make.

Further research shows the MACHTYPE is also set the architecture
type, so explicitly set it with the requested --arch.