Add an Inline signal and extend SignalProperty to use an Signal class.
The InlineSignal class emits its signals directly rather than placing
the signals on the event queue.  This is useful for situations where the
event queue is not operational (i.e. for signals emitted on startup and
shutdown of the event loop) or when the signals need to be processed before
code can be run.