Remove the need for callbacks from attr and make jobs aware if PID.
Rewrite the port attr function to no longer require callbacks.  The port's
attributes can now be fetched using a Signal.  While here make the config
port stage fetch the port's attributes (if needed) directly, since it does
have a load of 1 (and does not need the attr_queue).

Make build_stage() return the make object if one is being employed.  The
return status is now: False (failed to start stage), True (stage completed),
None (stage in progress), Popen (stage in progress using that subprocess).
There is one exception, if the stage refetches the ports attribute, that
subprocess will not be visible.

Make all jobs set the PID if a subprocess is spawned to do the work.  One
exception is detailed above.