Use list2cmdline and improve performance of Popen.
Use the list2cmdline from subprocess to display the commands that we
would have executed.

Popen waits for a response from the subprocess to determine if a problem
occured for the child.  This consumes considerable resources (about 50%
when in noop mode).  Portbuilder should handle any instance where the
subprocess does not execute, also that code path is will tested.