Add base class for Builders, introduce signally and succeeded ports.
The Builder class provides the common code across all ports and should be
sufficient for consumers of the various builders.

The succeeded Builder variable keeps a record of all ports that have succeeded
with their primary stage (added using builder(port)).

The builders now emit signals to update changes in their structure.  i.e. a port
moves from pending to queued (to active).  To accommodated this the jobs were
extended to emit signals when the job starts.