Multiple fixes to building logic.
Track which dependencies have already been loaded and immediately initiate a
signal and allow externally initiated ports to be registered as a dependency
loading.  The former change resolves multiple stage signals from builders and
the latter change prevents conflicting and multiple builds of a port.

Refactor the adding building code for DependBuilder.

Steal ownership of port if a stage supersedes a previous owner and only emit
signals SUCCEEDED or DONE signal if we have not failed (and there are multiple
ways to fail).