Fix build logic.
Correctly emit signal for finished dependency (typo).

Do not try any dependency resolution method if the port failed at or before the
DEPEND stage.

Correctly call the package builder.  If the builder does not require the
previous stage then also adding the port to that stage causes the signals to be
emitted in the reverse order (install signals after package, causing a fault in
the logic).