Refactor and fix the stage build checks.
In _resolve_depend(), directly call _port_failed() instead of delaying the
propagation of the failure.  This means that _port_ready() no longer needs to
check for that case and allows it to be more aggressive with it's asserts.

Previously a port's dependent object was not called with status_changed() if
the port did not need to be built and was being resolved using the package
builder.

Refactor the checking code, this eliminates the special case for PACKAGE stage
(that stage is handled with a subclass of StageBuilder) and, since _port_build()
is called from only one location, inline _port_build().