Refactor environment variables from make.conf and build target handling.
Allow extra values to be specified when initialising the environment variables
based on make.conf.  This allows an external setup program to setup state based
on the contents of make.conf.

Remove the use of flags["package"] and add flags["target"].  The target flag
corresponds to the DEPENDS_TARGET value.  This will facilitate better mimicking
of the port infrastructure.

While here change flags["pkg_sys"] to flags["pkg_mgmt"] to reflect that the
flag controls which package manager is used.