Refactor mk.status() into the ``pkg'' module.
This is in preparation for introducing support for pkgng.

Instead of inspecting the PKG_DBDIR directly call pkg.info, which will rely on
the public API (via `pkg_info').

Fix a bug where the distfile was incorrectly joined with the flags["chroot"]
path.  os.path.join ignores the previous parameters if one starts with '/', so
flags["chroot"] needs to be added before being passed to os.path.join.