Conditionally clean ports pre-build and post-build.
Extend Port.clean() to return the job used to clean and to allow forcing the
clean in-case an existing dirty (but unknown) state needs to be cleaned.

Once a port is added to the build stage queue it for building if pre-cleaning is
required.  The cleaning will be done in parallel to other stages and will only
delay the build stage, if needed.