Fix cyclic imports.
In libpb:
 1) pkg -> port.port -> pkg
 2) env -> port.port -> pkg -> env

To break (1) pkg no longer requires install status flags from port.port.Port,
these flags need to be moved into this module (as this is where the pkgdb is
located, which is the source of this flag).

To break (2) use the values for the install status flags, not the named values.
Ideally the named values from pkg should be used except pkg imports env which
would create another cyclic import.