Refactor the Builder._port_check() method and properly support "recursive" mode.
With the change to Dependent() Builder() now needs to handle "recursive" mode,
as such it needs to load all the dependencies directly.  Also, if the port is
not to be built at this stage then do not add it to the builder for the previous
stage.  This is a partial fix for 0.1.3.1-0 as build dependencies will not be
loaded, however it will load library and run dependencies.

Change the Builder._port_check() method to not have any side-affects.  Also,
due to the changes made to Dependent() there is no longer a need to issue a
port.dependent.status_changed().