Although pkgng supports chroot one cannot assume it is available in non-chroot.
The pkgng tools support direct operation within a chroot environment however
one cannot assume that it is installed in the non-chroot environment or that
if it is present that it is compatible with the chroot environment (i.e.
different versions).

Although one could install or upgrade the non-chroot pkgng tools it is not
justifiable to change the non-chroot environment as one would expect the '-C'
option to not mess around with the non-chroot environment.