Split load_defaults() to bootstrapping env.master and general options detection.
The original values for env.master are defaults that can be changed via either
os.environ or make.conf, if that is the case those changed values need to be
taken as the "original" values.

Create a function bootstrap_master() to detect the default values of env.master.

With the correct default values available the load_defaults() function can load
the bsd.ports.mk state variables directly from bsd.ports.mk and not from
make.conf.  This allows handling of dynamic handling of variables, such as
WITH_PKGNG, that is set if ${OSVERSION} > 1000017.