Merge mk.cache() into mk.load_defaults().
Use the same tactic with loading state variables for loading cached variables,
namely, use bsd.ports.mk to determine the correct values, avoiding the need to
track changes to the logic used.

There are many more variables in $PORTSDIR/Mk that could be cached however that
is left as a junior task.