Refactor Event() and add AdeptFuncEvent().
Make Event() a purely abstract base class with FuncEvent() replacing
Event()'s capability and create AdaptFuncEvent() that only passes the
arguments supported by the receiving function.

The intended use for AdaptFuncEvent() is for Signal()s to provide a
loser coupling with slots.  The processing of the arguments are done on
dispatch() to save space (args and kwargs) and to raise TypeError() in a
consistent mannor to FuncEvent().